Rename and move asset conversion helpers.
